Backend/requirements.txt
fastapi
uvicorn[standard]   
pandas
numpy
python-multipart
matplotlib
seaborn
scikit-learn
groq
python-dotenv
chardet
google-cloud-firestore
python-dotenv